consoliated launch.py and wsgi.py into __main__.py
